add functionality for rendering tips based on user's selected category checkboxes
