+++ results/index.html
@@ -32,7 +32,7 @@
                 <hr />
                 
             </div>
-            <div id="food-plan">
+            <!-- <div id="food-plan">
                 <h3>Food</h3>
                 <hr />
                 
@@ -61,7 +61,7 @@
                 <h3>Sanitation</h3>
                 <hr />
                 
-            </div>
+            </div> -->
     </main>
     <script type="module" src="results.js"></script>
 </body>
