+++ results/render-category-div.js
@@ -1,29 +1,33 @@
+import { renderCheckedOptionsTips } from './render-checked-options-tips.js';
+import { findById } from '../common/utils.js';
 
-export function renderCategoryDiv() {
+export function renderCategoryDiv(category, categoryId, user) {
     const containerDiv = document.querySelector('.container');
 
     const newDiv = document.createElement('div');
     const h3 = document.createElement('h3');
     h3.textContent = category.title;
     newDiv.appendChild(h3);
-    
+
+    // loop through selected options and render tips content
+    // renderCheckedOptionsTips();
+    for (let i = 0; i < user.responses[categoryId].length; i++) {
+        const userOption = user.responses[categoryId][i];
+
+        const userOptionObject = findById(category.options, userOption);
+        const label = renderCheckedOptionsTips(userOptionObject);
+
+        newDiv.appendChild(label);
+    }
+
+    const hr = document.createElement('hr');
+    newDiv.appendChild(hr);
+
+    containerDiv.appendChild(newDiv);
+
+
 }  
 
-function x() {
-    
-    const label = document.createElement('label');
-    const input = document.createElement('input');
-    input.type = 'checkbox';
-    label.appendChild(input);
-    const span = document.createElement('span');
-    span.textContent = category.options[];
-    label.appendChild(span);
-    const pTag = document.createElement('p');
-    pTag.textContent = category.options[]
-    
-}
 
-
-    newDiv.appendChild(label);
 
     //
